dheera_v0.3.0/dheera_v3/requirements.txt
# =====================================================
# Dheera v0.3.0 - Python Dependencies
# =====================================================

# Deep Learning Framework
torch>=2.0.0

# Numerical Computing
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# RAG & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# HTTP Requests (for Ollama, web search, etc.)
requests>=2.31.0

# Optional: Additional NLP tools
# transformers>=4.30.0  # Uncomment if using Hugging Face models directly

# Optional: OpenAI/Anthropic API clients (if not using Ollama)
# openai>=1.0.0
# anthropic>=0.3.0

# Development & Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
